# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi Demo Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/replays/{id}/getById": {
            "get": {
                "tags": [
                    "Replays"
                ],
                "summary": "Get single replays",
                "description": "Returns Get single replays",
                "operationId": "GetSinglereplay",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "replays id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Replay"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/replays/create": {
            "post": {
                "tags": [
                    "Replays"
                ],
                "summary": "Insert new Replay",
                "description": "Returns Replay data",
                "operationId": "Insert Replay",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Replay"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Replay"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/replays/{id}/update": {
            "post": {
                "tags": [
                    "Replays"
                ],
                "summary": "Insert new Replay info",
                "description": "Returns Replay data",
                "operationId": "Update Replay",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Replay id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Replay"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Replay"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/replays/{id}/delete": {
            "get": {
                "tags": [
                    "Replays"
                ],
                "summary": "Delete single Replay",
                "description": "Returns Deleted",
                "operationId": "DeleteSingleReplay",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Replay id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Replay"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
...
        "/Users/{id}/getById": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get single Users",
                "description": "Returns Get single Users",
                "operationId": "GetSingleUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Users id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/{id}/delete": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete single User",
                "description": "Returns Deleted",
                "operationId": "DeleteSingleUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/user/getOTP": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "getOTP",
                "description": "Returns User otp",
                "operationId": "getOTP",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/user/loginUsingOTP": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "loginUsingOTP",
                "description": "Returns Access otp",
                "operationId": "loginUsingOTP",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/User"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Replay": {
                "title": "Replay",
                "description": "Replay model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "replay": {
                        "title": "replay",
                        "description": "replay body",
                        "type": "string",
                        "example": "I can do it"
                    },
                    "images": {
                        "title": "images",
                        "description": "photo of a model",
                        "type": "string",
                        "example": "photo"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Replay"
                }
            },
...
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "name",
                        "description": "name of the new user",
                        "type": "string",
                        "example": "sara"
                    },
                    "role": {
                        "title": "role",
                        "description": "role of the new user",
                        "type": "string",
                        "example": "admin"
                    },
                    "phone": {
                        "title": "phone",
                        "description": "phone of the new user",
                        "type": "string",
                        "example": "07707722504"
                    },
                    "password": {
                        "title": "password",
                        "description": "Password of the new user",
                        "type": "string",
                        "example": "sara1234"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User"
                }
            }
        },
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}